% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{asm_paracou_6}
\alias{asm_paracou_6}
\title{Paracou plot 6}
\format{
A \link[tibble:tibble]{tibble::tibble}, which is a \link{data.frame}.
}
\source{
Permanent data census of Paracou: \url{https://paracou.cirad.fr/}
}
\usage{
asm_paracou_6
}
\description{
A community assembly, which is a tibble (and a data.frame).
It contains number of trees per species of the plot #6 of Paracou.
The plot covers 6.25 ha of tropical rainforest, divided into 4 equal-size subplots.
Each line of the tibble is a subplot.
The "site" column contains the subplot number, "weight" contains its area and all others columns contain a species.
Data are the number of trees above 10 cm diameter at breast height (DBH).
}
\details{
This dataset is from Paracou field station, French Guiana, managed by \href{https://www.cirad.fr}{Cirad}.
}
\keyword{datasets}
