% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ent_accum.R
\name{div_accum}
\alias{div_accum}
\alias{ent_accum}
\alias{ent_accum.numeric}
\alias{ent_accum.abundances}
\alias{div_accum.numeric}
\alias{div_accum.abundances}
\alias{autoplot.accumulation}
\title{Diversity Accumulation of a Community}
\usage{
ent_accum(x, ...)

\method{ent_accum}{numeric}(
  x,
  q = 0,
  levels = seq_len(sum(x)),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  jack_alpha = 0.05,
  jack_max = 10,
  n_simulations = 0,
  alpha = 0.05,
  show_progress = TRUE,
  ...,
  check_arguments = TRUE
)

\method{ent_accum}{abundances}(
  x,
  q = 0,
  levels = NULL,
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  jack_alpha = 0.05,
  jack_max = 10,
  n_simulations = 0,
  alpha = 0.05,
  show_progress = TRUE,
  ...,
  check_arguments = TRUE
)

div_accum(x, ...)

\method{div_accum}{numeric}(
  x,
  q = 0,
  levels = seq_len(sum(x)),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  jack_alpha = 0.05,
  jack_max = 10,
  n_simulations = 0,
  alpha = 0.05,
  show_progress = TRUE,
  ...,
  check_arguments = TRUE
)

\method{div_accum}{abundances}(
  x,
  q = 0,
  levels = NULL,
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  jack_alpha = 0.05,
  jack_max = 10,
  n_simulations = 0,
  alpha = 0.05,
  show_progress = TRUE,
  ...,
  check_arguments = TRUE
)

\method{autoplot}{accumulation}(
  object,
  ...,
  main = NULL,
  xlab = "Sample Size",
  ylab = NULL,
  shade_color = "grey75",
  alpha = 0.3,
  border_color = "red",
  lty = ggplot2::GeomLine$default_aes$linetype,
  lwd = ggplot2::GeomLine$default_aes$linewidth
)
}
\arguments{
\item{x}{An object, that may be a numeric vector containing abundances or probabilities,
or an object of class \link{abundances}  or \link{probabilities}}

\item{...}{Unused.}

\item{probability_estimator}{One of the estimators of a probability distribution:
"naive" (the default value), or "Chao2013", "Chao2015", "ChaoShen" to estimate
the probabilities of the observed species in the asymptotic distribution.}

\item{unveiling}{One of the possible unveiling methods to estimate the probabilities
of the unobserved species: "none" (default, no species is added), "uniform"
(all unobserved species have the same probability) or "geometric" (the
unobserved species distribution is geometric).}

\item{richness_estimator}{An estimator of richness to evaluate the total number of species,
see \link{div_richness}.}

\item{jack_alpha}{The risk level, 5\% by default, used to optimize the jackknife order.}

\item{jack_max}{The highest jackknife order allowed. Default is 10.}

\item{alpha}{The opacity of the confidence envelopes, between 0 (transparent) and 1 (opaque).}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{object}{An object of class "accumulation".}

\item{main}{The main title of the plot.}

\item{xlab}{The label of the x-axis.}

\item{ylab}{The label of the y-axis.}

\item{shade_color}{The color of the shaded confidence envelopes.}

\item{border_color}{The color of the border of the confidence envelopes.}

\item{lty}{The line type of the curves.}

\item{lwd}{The line width of the curves.}

\item{estimator}{An estimator of entropy.}

\item{level}{The level of interpolation or extrapolation.
It may be a chosen sample size (an integer) or a sample coverage
(a number between 0 and 1).
Richness extrapolation require its asymptotic estimation depending on the
choice of the \code{estimator}.}
}
\value{
A tibble with the site names, the estimators used and the accumulated entropy
or diversity at each level of sampling effort.
}
\description{
Details TODO.
}
\examples{
# TODO

}
\references{
\insertAllCited{}
}
