% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_similarity.R
\name{div_similarity}
\alias{div_similarity}
\alias{div_similarity.numeric}
\alias{div_similarity.species_distribution}
\title{Similarity-Based Diversity of a Community}
\usage{
div_similarity(x, similarities, q = 1, ...)

\method{div_similarity}{numeric}(
  x,
  similarities = diag(length(x)),
  q = 1,
  estimator = c("UnveilJ", "Max", "ChaoShen", "MarconZhang", "UnveilC", "UnveiliC",
    "naive"),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  jack_alpha = 0.05,
  jack_max = 10,
  sample_coverage = NULL,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  as_numeric = FALSE,
  ...,
  check_arguments = TRUE
)

\method{div_similarity}{species_distribution}(
  x,
  similarities = diag(sum(!colnames(x) \%in\% non_species_columns)),
  q = 1,
  estimator = c("UnveilJ", "Max", "ChaoShen", "MarconZhang", "UnveilC", "UnveiliC",
    "naive"),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  jack_alpha = 0.05,
  jack_max = 10,
  gamma = FALSE,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  ...,
  check_arguments = TRUE
)
}
\arguments{
\item{x}{An object, that may be a numeric vector containing abundances or probabilities,
or an object of class \link{abundances} or \link{probabilities}.
If it is a numeric vector, then its length must equal the dimensions of the
\code{similarities} matrix: species are assumed to be in the same order.}

\item{similarities}{A similarity matrix, that can be obtained by \link{fun_similarity}.
Its default value is the identity matrix.}

\item{q}{The order of diversity.}

\item{...}{Unused.}

\item{estimator}{An estimator of asymptotic diversity.}

\item{probability_estimator}{A string containing one of the possible estimators
of the probability distribution (see \link{probabilities}).
Used only for extrapolation.}

\item{unveiling}{A string containing one of the possible unveiling methods
to estimate the probabilities of the unobserved species (see \link{probabilities}).
Used only for extrapolation.}

\item{jack_alpha}{The risk level, 5\% by default, used to optimize the jackknife order.}

\item{jack_max}{The highest jackknife order allowed. Default is 10.}

\item{sample_coverage}{The sample coverage of \code{x} calculated elsewhere.
Used to calculate the gamma diversity of meta-communities, see details.}

\item{coverage_estimator}{An estimator of sample coverage used by \link{coverage}.}

\item{as_numeric}{If \code{TRUE}, a number or a numeric vector is returned rather than a tibble.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{gamma}{If \code{TRUE}, \eqn{\gamma} diversity, i.e. diversity of the metacommunity, is computed.}
}
\value{
A tibble with the site names, the estimators used and the estimated diversity.
}
\description{
Estimate the diversity of species from abundance or probability data and a
similarity matrix between species.
Several estimators are available to deal with incomplete sampling.
Bias correction requires the number of individuals.
}
\details{
All species of the \code{species_distribution} must be found in the matrix of
\code{similarities} if it is named.
If it is not, its size must equal the number of species.
Then, the order of species is assumed to be the same as that of the
\code{species_distribution}.

Similarity-Based diversity can't be interpolated of extrapolated as of the
state of the art.
}
\examples{
# Similarity matrix
Z <- fun_similarity(paracou_6_fundist)
# Diversity of each community
div_similarity(paracou_6_abd, similarities = Z, q = 2)
# gamma diversity
div_similarity(paracou_6_abd, similarities = Z, q = 2, gamma = TRUE)

}
\references{
\insertAllCited{}
}
