% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{divent}
\alias{divent}
\title{divent}
\description{
Measures of Diversity and Entropy
}
\details{
This package is a reboot of the \strong{entropart} package \insertCite{Marcon2014c}{divent}.
}
\references{
\insertAllCited{}
}
