% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq.R
\name{e_n_q}
\alias{e_n_q}
\title{Grassberger's expectation of n^q}
\usage{
e_n_q(n, q)
}
\arguments{
\item{n}{A positive integer vector.}

\item{q}{A positive number.}
}
\value{
A vector of the same length as n containing the transformed values.
}
\description{
Expected value of \eqn{n^q} when \eqn{n} follows a Poisson distribution
of parameter \eqn{n}.
}
\details{
The expectation of \eqn{n^q} when \eqn{n} follows a Poisson distribution
was derived by \insertCite{Grassberger1988;textual}{divent}.

The function is computed using the \link{beta} function.
Its value is 0 for \eqn{n-q+1<0}.
}
\examples{
# Compare
n <- c(2, 3)
e_n_q(n, q = 2)
# with
n^2

# Result is 1
e_n_q(n, q = 0)
# Result is 0
e_n_q(n, q=5)
}
\references{
\insertAllCited{}
}
