% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ordinariness.R
\name{fun_ordinariness}
\alias{fun_ordinariness}
\title{Functional ordinariness of a community}
\usage{
fun_ordinariness(
  distribution,
  similarities,
  as_numeric = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{distribution}{TODO}

\item{similarities}{TODO}

\item{as_numeric}{If \code{TRUE}, a number or a numeric vector is returned rather than a tibble.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\value{
A tibble with the ordinariness of each species.
}
\description{
TODO
}
\examples{
fun_ordinariness(paracou_6_abd, fun_similarity(paracou_6_fundist))

}
