% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{non_species_columns}
\alias{non_species_columns}
\title{Non-Species columns}
\format{
A character vector.
}
\usage{
non_species_columns
}
\description{
The column names that are not those of species in a \link{species_distribution}.
}
\details{
Objects of classes \link{abundances} and \link{probabilities}, that are also of class
\link{species_distribution}, have columns named after the species they contain.
Some columns are reserved to describe the plots and their diversity.
}
\keyword{datasets}
