% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{paracou_6}
\alias{paracou_6}
\alias{paracou_6_abd}
\alias{paracou_6_wmppp}
\title{Paracou plot 6}
\format{
\code{paracou_6_abd} is an object of class \link{abundances}, which is also a \link[tibble:tibble]{tibble::tibble}.
Each line of the tibble is a subplot.
\code{paracou_6_wmppp} is a \link[dbmss:wmppp]{dbmss::wmppp} object, i.e. a weighted, marked planar point pattern.

An object of class \code{wmppp} (inherits from \code{ppp}) of length 6.
}
\source{
Permanent data census of Paracou: \url{https://paracou.cirad.fr/}
}
\usage{
paracou_6_abd

paracou_6_wmppp
}
\description{
A community assembly.
It contains number of trees per species of the plot #6 of Paracou.
The plot covers 6.25 ha of tropical rainforest, divided into 4 equally-sized subplots.
}
\details{
In \code{paracou_6_abd} (a tibble), the "site" column contains the subplot number, "weight" contains its area and all others columns contain a species.
Data are the number of trees above 10 cm diameter at breast height (DBH).

In \code{paracou_6_wmppp} (a point pattern), the point type is tree species and the point weight is their basal area, in square centimeters.

This dataset is from Paracou field station, French Guiana, managed by \href{https://www.cirad.fr}{Cirad}.
}
\examples{
# Rank-abundance curve of the species of the whole plot
autoplot(metacommunity(paracou_6_abd))

}
\seealso{
\link{paracou_6_taxo}, \link{paracou_6_fundist}
}
\keyword{datasets}
