% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{paracou_6_fundist}
\alias{paracou_6_fundist}
\title{Functional distances between Paracou plot 6 species}
\format{
A matrix.
}
\source{
Permanent data census of Paracou: \url{https://paracou.cirad.fr/}
}
\usage{
paracou_6_fundist
}
\description{
A functional distance matrix of species of the dataset \link{paracou_6_abd}.
Distances were computed from a trait dataset including specific leaf area,
wood density, seed mass and 95th percentile of height. Gower's metric
\insertCite{Gower1971}{divent} was used to obtain a distance matrix.
}
\details{
This dataset is from Paracou field station, French Guiana, managed by \href{https://www.cirad.fr}{Cirad}.
}
\seealso{
\link{paracou_6_abd}, \link{paracou_6_taxo}
}
\keyword{datasets}
