% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{paracou_6_taxo}
\alias{paracou_6_taxo}
\title{Taxonomy of Paracou plot 6 species}
\format{
An object of class \link[ape:phylo]{ape::phylo}, which is a phylogenetic tree.
}
\source{
Permanent data census of Paracou: \url{https://paracou.cirad.fr/}
}
\usage{
paracou_6_taxo
}
\description{
The taxonomy of species of the dataset \link{paracou_6_abd}.
Distances in the tree are 1 (different species of the same genus),
2 (same family) or 3 (different families).
}
\details{
This dataset is from Paracou field station, French Guiana, managed by \href{https://www.cirad.fr}{Cirad}.
}
\seealso{
\link{paracou_6_abd}, \link{paracou_6_fundist}
}
\keyword{datasets}
