% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelCountNbd}
\alias{parallelCountNbd}
\title{parallelCountNbd}
\usage{
parallelCountNbd(r, NbSpecies, x, y, Type, Weight)
}
\arguments{
\item{r}{The vector of distances to take into account.}

\item{NbSpecies}{The number of species of the community.}

\item{x, y}{The coordinates of the points.}

\item{Type}{A vector containing the species of each point (as integer, i.e. the factor code).}

\item{Weight}{A vector containing the weights of points.}
}
\value{
a numeric vector with the distances to convert to a matrix
}
\description{
Create a 3-D array containing the number of neighbors around each point, per species
}
\examples{
X <- rspcommunity(1, size = 5, species_number = 3)
matrix(parallelCountNbd(r = 0:1, NbSpecies = 3, x= X$x, y = X$y, 
    Type = marks(X)$PointType, Weight = marks(X)$PointType), nrow = 5)
}
