% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map}
\usage{
plot_map(
  accum,
  q = dimnames(accum$Accumulation)$q[1],
  neighborhood = dplyr::last(colnames(accum$Neighborhoods)),
  sigma = spatstat.explore::bw.scott(accum$X, isotropic = TRUE),
  allow_jitter = TRUE,
  weighted = FALSE,
  adjust = 1,
  dim_x = 128,
  dim_y = 128,
  col = terrain.colors(256),
  contour = TRUE,
  contour_levels = 10,
  contour_col = "dark red",
  points = FALSE,
  pch = 20,
  point_col = "black",
  ...,
  check_arguments = TRUE
)
}
\arguments{
\item{accum}{an object to map.}

\item{q}{a number: the order of diversity.}

\item{neighborhood}{The neighborhood size, i.e. the number of neighbors or the distance to consider.}

\item{sigma}{the smoothing bandwidth.
The standard deviation of the isotropic smoothing kernel.
Either a numerical value, or a function that computes an appropriate value of sigma.}

\item{allow_jitter}{if \code{TRUE}, duplicated points are jittered to avoid their
elimination by the smoothing procedure.}

\item{weighted}{if \code{TRUE}, the weight of the points is used by the smoothing
procedure.}

\item{adjust}{force the automatically selected bandwidth to be multiplied
by \code{adjust}.
Setting it to values lower than one (1/2 for example) will sharpen the estimation.}

\item{dim_x}{the number of columns (pixels) of the resulting map, 128 by default.}

\item{dim_y}{the number of rows (pixels) of the resulting map, 128 by default.}

\item{col}{the colors of the map. See \link[spatstat.geom:plot.im]{spatstat.geom::plot.im} for details.}

\item{contour}{if \code{TRUE}, contours are added to the map.}

\item{contour_levels}{the number of levels of contours.}

\item{contour_col}{the color of the contour lines.}

\item{points}{if \code{TRUE}, the points that brought the data are added to the map.}

\item{pch}{the symbol used to represent points.}

\item{point_col}{The color of the points.}

\item{...}{further arguments passed to \link[spatstat.explore:bw.smoothppp]{spatstat.explore::bw.smoothppp} and
\link[spatstat.explore:density.ppp]{spatstat.explore::density.ppp} to control the kernel smoothing and
to \link[spatstat.geom:plot.im]{spatstat.geom::plot.im} to plot the image.
Standard base graphic arguments such as \code{main} can be used.}

\item{check_arguments}{if \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\value{
A \link[spatstat.geom:im]{spatstat.geom::im} object that can be used to produce
alternative maps.
}
\description{
Map Spatialized Communities
}
\details{
Maps of interpolated values from points of an \link{accum} produced e.g. by the function \link{DivAccum} object are produced.
}
\examples{
# Generate a random community
X <- rspcommunity(1, size = 50, species_number = 10)
# Calculate the species accumulation curve
accum <- accum_sp_hill(X, orders = 0, r = c(0, 0.2), individual=TRUE)
# Plot the local richness at distance = 0.2
plot_map(accum, q = 0, neighborhood = 0.2)

}
