% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{probabilities}
\alias{probabilities}
\title{Probabilities of Species}
\usage{
probabilities(x, ...)
}
\arguments{
\item{x}{An object. It may be:
\itemize{
\item a numeric vector containing abundances. It may be named to track species names.
\item an object of class \link{species_distribution}.
}}

\item{...}{Unused.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\description{
Estimate actual probabilities of species from a sample
}
\details{
\code{probabilities()} estimates a probability distribution from a sample.
If the \code{estimator} is not "naive", the observed abundance distribution is used
to estimate the actual probability distribution.
The list of species is changed:
zero-abundance species are cleared, and some unobserved species are added.
First, observed species probabilities are estimated following
\insertCite{Chao2003;textual}{divent}, i.e. input probabilities are multiplied by
the sample coverage, or according to more sophisticated models:
\insertCite{Chao2013;textual}{divent}, a single-parameter model, or
\insertCite{Chao2015;textual}{divent}, a two-parameter model.
The total probability of observed species equals the sample coverage.
Then, the distribution of unobserved species can be unveiled: their number
is estimated according to the \code{richness_estimator}.
The coverage deficit (1 minus the sample coverage) is shared by the unobserved
species equally (\code{unveiling} = "uniform", \insertCite{Chao2013}{divent}) or
according to a geometric distribution (\code{unveiling} = "geometric", \insertCite{Chao2015}{divent}).
}
\examples{
probabilities(paracou_6_abd)
}
\references{
\insertAllCited{}
}
