% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{sample_coverage}
\alias{sample_coverage}
\alias{coverage}
\alias{coverage.numeric}
\alias{coverage.abundances}
\alias{coverage_to_size}
\title{Sample Coverage of a Community}
\usage{
coverage(x, ...)

\method{coverage}{numeric}(
  x,
  estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  level = NULL,
  ...,
  check_arguments = TRUE
)

\method{coverage}{abundances}(
  x,
  estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  level = NULL,
  ...,
  check_arguments = TRUE
)

coverage_to_size(abundances, sample_coverage, check_arguments = TRUE)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}

\item{estimator}{A string containing one of the possible estimators:
"ZhangHuang", "Chao", "Turing", "Good".}

\item{level}{The level of interpolation or extrapolation, i.e. an abundance.}

\item{check_arguments}{If \code{TRUE}, check the arguments of the function.}

\item{abundances}{An object of class \code{abundances}.}

\item{sample_coverage}{The target sample coverage.}
}
\value{
\code{coverage()} returns a named number equal to the calculated sample coverage.
The name is that of the estimator used.
\code{coverage_to_size()} returns a number equal to the sample size corresponding
to the chosen sample coverage.
}
\description{
\code{coverage()} calculates an estimator of the sample coverage of a community
described by its abundance vector.
\code{coverage_to_size()} estimates the sample size corresponding to the chosen
sample coverage.
}
\examples{
coverage(asm_paracou_6)

}
\references{
\insertAllCited{}
}
