% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abd_species.R
\name{abd_species}
\alias{abd_species}
\alias{abd_sum}
\title{Abundances of Communities}
\usage{
abd_species(abundances, check_arguments = TRUE)

abd_sum(abundances, as_numeric = FALSE, check_arguments = TRUE)
}
\arguments{
\item{abundances}{An object of class of class \link{abundances}.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{as_numeric}{If \code{TRUE}, a number is returned rather than a tibble.}
}
\description{
Utilities for community abundances (objects of class "abundances").
}
\details{
\code{abd_species()} returns a tibble containing the species abundance columns only,
to simplify numeric operations.

\code{abd_sum()} returns the sample sizes of the communities.
}
\examples{
abd_species(paracou_6_abd)
abd_sum(paracou_6_abd)

}
