% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_sp_hill.R
\name{accum_sp_hill}
\alias{accum_sp_hill}
\alias{accum_sp_tsallis}
\alias{accum_mixing}
\title{Spatial Diversity Accumulation of a Community}
\usage{
accum_sp_tsallis(
  X,
  orders = 0,
  neighbors = 1:ceiling(X$n/2),
  r = NULL,
  correction = c("none", "extrapolation"),
  entropy_estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger",
    "Marcon", "UnveilC", "UnveiliC", "ZhangGrabchak", "naive", "Bonachela", "Holste"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  jack_alpha = 0.05,
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  individual = FALSE,
  show_progress = TRUE,
  check_arguments = TRUE
)

accum_sp_hill(
  X,
  orders = 0,
  neighbors = 1:ceiling(X$n/2),
  r = NULL,
  correction = c("none", "extrapolation"),
  entropy_estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger",
    "Marcon", "UnveilC", "UnveiliC", "ZhangGrabchak", "naive", "Bonachela", "Holste"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  jack_alpha = 0.05,
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  h0 = c("none", "multinomial", "random location", "binomial"),
  alpha = 0.05,
  n_simulations = 100,
  individual = FALSE,
  show_progress = TRUE,
  check_arguments = TRUE
)

accum_mixing(
  X,
  orders = 0,
  neighbors = 1:ceiling(X$n/2),
  r = NULL,
  correction = c("none", "extrapolation"),
  entropy_estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger",
    "Marcon", "UnveilC", "UnveiliC", "ZhangGrabchak", "naive", "Bonachela", "Holste"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  probability_estimator = c("Chao2015", "Chao2013", "ChaoShen", "naive"),
  unveiling = c("geometric", "uniform", "none"),
  jack_alpha = 0.05,
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  h0 = c("none", "multinomial", "random location", "binomial"),
  alpha = 0.05,
  n_simulations = 100,
  individual = FALSE,
  show_progress = TRUE,
  check_arguments = TRUE
)
}
\arguments{
\item{X}{a spatialized community
(A \link[dbmss:wmppp]{dbmss::wmppp} object with \code{PointType} values as species names.)}

\item{orders}{a numeric vector: the diversity orders to address. Default is 0.}

\item{neighbors}{a vector of integers.
Entropy will be accumulated along this number of neighbors around each individual.
Default is 10\% of the individuals.}

\item{r}{a vector of distances.
If \code{NULL} accumulation is along \code{n}, else neighbors are accumulated in circles of radius \code{r}.}

\item{correction}{The edge-effect correction to apply when estimating
the entropy of a neighborhood community that does not fit in the window.
Does not apply if neighborhoods are defined by the number of neighbors.
Default is "none".
"extrapolation" extrapolates the observed diversity up to the number of individuals
estimated in the full area of the neighborhood, which is slow.}

\item{entropy_estimator}{The asymptotic estimator of entropy
when \code{correction} is "extrapolation".}

\item{richness_estimator}{an estimator of richness to evaluate the total number of species,
see \link{div_richness}. used for interpolation and extrapolation.}

\item{probability_estimator}{a string containing one of the possible estimators
of the probability distribution (see \link{probabilities}).
Used only for extrapolation.}

\item{unveiling}{a string containing one of the possible unveiling methods
to estimate the probabilities of the unobserved species (see \link{probabilities}).
Used only for extrapolation.}

\item{jack_alpha}{the risk level, 5\% by default, used to optimize the jackknife order.}

\item{jack_max}{the highest jackknife order allowed. Default is 10.}

\item{coverage_estimator}{an estimator of sample coverage used by \link{coverage}.}

\item{individual}{If \code{TRUE}, individual neighborhood entropies are returned.}

\item{show_progress}{if TRUE, a progress bar is shown during long computations.}

\item{check_arguments}{if \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{h0}{The null hypothesis to compare the distribution of \code{X} to.
If "none", the default value, no null hypothesis is tested.
"multinomial" means the community will be rarefied down to the number of \code{neighbors}.
"random location" means the points will we randomly permuted across their actual locations.
"binomial" means the points will we uniformly and independently drawn
in the window (a binomial point process is a Poisson point process conditionally to the number of points).}

\item{alpha}{the risk level, 5\% by default.}

\item{n_simulations}{the number of simulations used to estimate the confidence envelope.}
}
\value{
an \link{accum_sp} object, that is also either an \link{accum_sp_diversity},
\link{accum_sp_entropy} or \link{accum_sp_mixing} object.
}
\description{
Spatial Diversity and Entropy Accumulation Curves represent the accumulation of
entropy and diversity with respect to the distance from individuals
}
\details{
\code{accum_sp_hill()} or \code{accum_sp_tsallis()} estimate the diversity or entropy
accumulation curve of a distribution.

These functions can compute the entropy of the neighborhood of points in parallel
thanks to package \link[doFuture:doFuture]{doFuture}.
A parallel plan must be chosen with function \code{\link[future:plan]{future::plan()}}.
To use all cores but one on a personal computer, run:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library("doFuture")
plan(multisession, workers = availableCores(omit = 1))
}\if{html}{\out{</div>}}

If argument \code{show_progress} is \code{TRUE}, a progress bar can be displayed
by package \link[progressr:progressr]{progressr}.
Since all progress bars rely on package \link[cli:cli]{cli} in \emph{divent}, the "cli" handler
is a good choice.

\if{html}{\out{<div class="sourceCode">}}\preformatted{library("progressr")
handlers(global = TRUE)
handlers("cli")
}\if{html}{\out{</div>}}

Those code chunks must be run before \code{accum_sp_hill()} or \code{accum_sp_tsallis()}
with large point patterns such as \link{paracou_6_wmppp}.

Argument \code{entropy_estimator} is used to estimate asymptotic entropy or diversity
when argument \code{correction} is "extrapolation".
Then, \code{richness_estimator} is used if needed to estimate entropy.
If the jackknife is chosen, \code{jack_alpha} sets the uncertainty level that decides
increasing the jackknife estimator order, and \code{jack_max} is the highest
order allowed.
If \code{entropy_estimator} is an unveiled one, then \code{probability_estimator} and
\code{unveiling} define the method.
\code{coverage_estimator} sets the estimator of sample coverage.
In general, the default values of these arguments are a good choice.
}
\examples{
# Generate a random community
X <- rspcommunity(1, size = 50, species_number = 3)
# Calculate the accumulation of richness
accum_n <- accum_sp_hill(X)
plot(accum_n, q = 0)
# along distance
accum_r <- accum_sp_hill(X, orders = 1, r = seq(0, .5, .05))
autoplot(accum_r, q = 1)

}
