% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_sp_hill.R
\name{accum_sp_hill}
\alias{accum_sp_hill}
\alias{accum_sp_tsallis}
\alias{accum_mixing}
\alias{plot.Accumulation}
\alias{autoplot.Accumulation}
\title{Spatial Diversity Accumulation of a Community}
\usage{
accum_sp_tsallis(
  X,
  orders = 0,
  estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger", "Marcon",
    "UnveilC", "UnveiliC", "ZhangGrabchak", "naive", "Bonachela", "Holste"),
  neighbors = 1:ceiling(X$n/2),
  r = NULL,
  correction = c("none", "extrapolation"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  individual = FALSE,
  show_progress = TRUE,
  check_arguments = TRUE
)

accum_sp_hill(
  X,
  orders = 0,
  estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger", "Marcon",
    "UnveilC", "UnveiliC", "ZhangGrabchak", "naive", "Bonachela", "Holste"),
  neighbors = 1:ceiling(X$n/2),
  r = NULL,
  correction = c("none", "extrapolation"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  h0 = c("none", "multinomial", "random location", "binomial"),
  alpha = 0.05,
  n_simulations = 100,
  individual = FALSE,
  show_progress = TRUE,
  check_arguments = TRUE
)

accum_mixing(
  X,
  orders = 0,
  estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger", "Marcon",
    "UnveilC", "UnveiliC", "ZhangGrabchak", "naive", "Bonachela", "Holste"),
  neighbors = 1:ceiling(X$n/2),
  r = NULL,
  correction = c("none", "extrapolation"),
  richness_estimator = c("rarefy", "jackknife", "iChao1", "Chao1", "naive"),
  h0 = c("none", "multinomial", "random location", "binomial"),
  alpha = 0.05,
  n_simulations = 100,
  individual = FALSE,
  show_progress = TRUE,
  check_arguments = TRUE
)

\method{plot}{Accumulation}(
  x,
  ...,
  q = dimnames(x$Accumulation)$q[1],
  type = "l",
  main = "Accumulation of ...",
  xlab = "Sample size...",
  ylab = "Diversity...",
  ylim = NULL,
  show_h0 = TRUE,
  line_width = 2,
  col_shade = "grey75",
  col_border = "red"
)

\method{autoplot}{Accumulation}(
  object,
  ...,
  q = dimnames(object$Accumulation)$q[1],
  main = "Accumulation of ...",
  xlab = "Sample size...",
  ylab = "Diversity...",
  ylim = NULL,
  show_h0 = TRUE,
  col_shade = "grey75",
  col_border = "red"
)
}
\arguments{
\item{X}{a spatialized community
(A \link[dbmss:wmppp]{dbmss::wmppp} object with \code{PointType} values as species names.)}

\item{orders}{A numeric vector: the diversity orders to address. Default is 0.}

\item{estimator}{an estimator of asymptotic entropy, diversity or richness.}

\item{neighbors}{A vector of integers.
Entropy will be accumulated along this number of neighbors around each individual.
Default is 10\% of the individuals.}

\item{r}{A vector of distances.
If \code{NULL} accumulation is along \code{n}, else neighbors are accumulated in circles of radius \code{r}.}

\item{correction}{The edge-effect correction to apply when estimating
the entropy of a neighborhood community that does not fit in the window.
Does not apply if neighborhoods are defined by the number of neighbors.
Default is "none".
"extrapolation" extrapolates the observed diversity up to the number of individuals
estimated in the full area of the neighborhood, which is slow.}

\item{richness_estimator}{an estimator of richness to evaluate the total number of species,
see \link{div_richness}. used for interpolation and extrapolation.}

\item{individual}{If \code{TRUE}, individual neighborhood entropies are returned.}

\item{show_progress}{if TRUE, a progress bar is shown during long computations.}

\item{check_arguments}{if \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{h0}{The null hypothesis to compare the distribution of \code{X} to.
If "none", the default value, no null hypothesis is tested.
"multinomial" means the community will be rarefied down to the number of \code{neighbors}.
"random location" means the points will we randomly permuted accross their actual locations.
"binomial" means the points will we uniformly and independently drawn
in the window (a binomial point process is a Poisson point process conditionally to the number of points).}

\item{alpha}{the risk level, 5\% by default.}

\item{n_simulations}{the number of simulations used to estimate the confidence envelope.}

\item{x}{an "Accumulation" object that can be accumulation of diversity (\code{\link{DivAccum}}), entropy (\code{\link{EntAccum}}) or the Mixing index (\code{\link{Mixing}}).}

\item{q}{a number: the order of diversity.}

\item{type}{plotting parameter. Default is "l".}

\item{main}{main title of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{ylim}{limits of the Y-axis, as a vector of two numeric values.}

\item{show_h0}{if \code{TRUE}, the values of the null hypothesis are plotted.}

\item{line_width}{width of the Diversity Accumulation Curve line.}

\item{col_shade}{The color of the shaded confidence envelope.}

\item{col_border}{The color of the borders of the confidence envelope.}

\item{object}{An "Accumulation" object that cat be accumulation of
diversity (\link{DivAccum}), entropy (\link{EntAccum}) or the Mixing index (\link{Mixing}).}
}
\value{
An "Accumulation" object that is a list.
\itemize{
\item Its first item, named "X", is \code{X}.
\item Its second item, named "Accumulation", is a 3-D array containing average entropy.
The third dimension of the array is only of length 1: it contains observed entropy.
The first two dimensions are respectively for $q$ values and the number of points
of the neighborhood, starting from 1 (the point itself, with no neighbor), or the distances starting from 0.
\item Its third item, named "Neighborhoods" has the same structure as the second one
but its third dimension contains the local values accumulated in the neighborhood of each point.
}
}
\description{
Spatial Diversity and Entropy Accumulation Curves represent the accumulation of
entropy and diversity with respect to the distance from individuals
}
\details{
\code{accum_sp_hill()} or \code{accum_sp_tsallis()} estimate the diversity or entropy
accumulation curve of a distribution.
}
\examples{
# Generate a random community
X <- rspcommunity(1, size = 50, species_number = 3)
# Calculate the accumulation of richness 
accum <- accum_sp_hill(X)
plot(accum, q = 0)
# along distance
accum_r <- accum_sp_hill(X, orders = 1, r = seq(0, .5, .05))
autoplot(accumR, q=1)

}
\references{
\insertAllCited{}
}
