% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{check_divent_args}
\alias{check_divent_args}
\title{check_divent_args}
\usage{
check_divent_args()
}
\value{
Returns \code{TRUE} or stops if a problem is detected.
}
\description{
Checks the arguments of a function of the package divent
}
\details{
The function compares the arguments passed to its parent function to the type
they should be and performs some extra tests, \emph{e.g.} probabilities must be positive and sum to 1.
It stops if an argument is not correct.
}
\author{
Eric Marcon \href{mailto:Eric.Marcon@agroparistech.fr}{Eric.Marcon@agroparistech.fr}
}
\keyword{internal}
