% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_hill.R
\name{div_hill}
\alias{div_hill}
\alias{div_hill.numeric}
\alias{div_hill.abundances}
\title{Hill number of a Community}
\usage{
div_hill(x, q = 1, ...)

\method{div_hill}{numeric}(
  x,
  q = 1,
  estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger", "Holste",
    "Marcon", "UnveilC", "UnveiliC", "ZhangGrabchak", "naive"),
  level = NULL,
  probability_estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "naive"),
  jack_alpha = 0.05,
  jack_max = 10,
  sample_coverage = NULL,
  as_numeric = FALSE,
  ...,
  check_arguments = TRUE
)

\method{div_hill}{abundances}(
  x,
  q = 1,
  estimator = c("UnveilJ", "ChaoJost", "ChaoShen", "GenCov", "Grassberger", "Holste",
    "Marcon", "UnveilC", "UnveiliC", "ZhangGrabchak"),
  level = NULL,
  probability_estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "naive"),
  jack_alpha = 0.05,
  jack_max = 10,
  ...,
  check_arguments = TRUE
)
}
\arguments{
\item{x}{An object, that may be a numeric vector containing abundances or probabilities,
or an object of class \link{abundances}  or \link{probabilities}}

\item{q}{The order or diversity.}

\item{...}{Unused.}

\item{estimator}{An estimator of entropy.}

\item{level}{The level of interpolation or extrapolation.
It may be a chosen sample size (an integer) or a sample coverage
(a number between 0 and 1).
Richness extrapolation require its asymptotic estimation depending on the
choice of the \code{estimator}.}

\item{probability_estimator}{One of the estimators of a probability distribution:
"naive" (the default value), or "Chao2013", "Chao2015", "ChaoShen" to estimate
the probabilities of the observed species in the asymptotic distribution.}

\item{unveiling}{One of the possible unveiling methods to estimate the probabilities
of the unobserved species: "none" (default, no species is added), "uniform"
(all unobserved species have the same probability) or "geometric" (the
unobserved species distribution is geometric).}

\item{richness_estimator}{An estimator of richness to evaluate the total number of species,
see \link{div_richness}.}

\item{jack_alpha}{The risk level, 5\% by default, used to optimize the jackknife order.}

\item{jack_max}{The highest jackknife order allowed. Default is 10.}

\item{sample_coverage}{The sample coverage of \code{x} calculated elsewhere.
Used to calculate the gamma diversity of meta-communities, see details.}

\item{as_numeric}{If \code{TRUE}, a number is returned rather than a tibble.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\value{
A tibble with the site names, the estimators used and the estimated diversity.
}
\description{
Estimate the diversity sensu stricto, i.e. the Hill number of species from abundance or probability data.
Several estimators are available to deal with incomplete sampling.
}
\details{
Bias correction requires the number of individuals.
}
\examples{
div_hill(paracou_6_abd, q = 2)

}
\references{
\insertAllCited{}
}
