% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{divent-package}
\alias{divent}
\alias{divent-package}
\title{divent}
\description{
Measures of Diversity and Entropy
}
\details{
This package is a reboot of the \strong{entropart} package \insertCite{Marcon2014c}{divent}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EricMarcon/divent}
  \item Report bugs at \url{https://github.com/EricMarcon/divent/issues}
}

}
\author{
\strong{Maintainer}: Eric Marcon \email{eric.marcon@agroparistech.fr} (\href{https://orcid.org/0000-0002-5249-321X}{ORCID})

}
