% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rac.R
\name{fit_rac}
\alias{fit_rac}
\alias{fit_rac.numeric}
\alias{fit_rac.species_distribution}
\title{Fit a distribution}
\usage{
fit_rac(x, ...)

\method{fit_rac}{numeric}(
  x,
  distribution = c("lnorm", "lseries", "geom", "bstick"),
  ...,
  check_arguments = TRUE
)

\method{fit_rac}{species_distribution}(
  x,
  distribution = c("lnorm", "lseries", "geom", "bstick"),
  ...,
  check_arguments = TRUE
)
}
\arguments{
\item{x}{An object}

\item{...}{Unused.}

\item{distribution}{The distribution of species abundances.
May be "lnorm" (log-normal), "lseries" (log-series), "geom" (geometric) or
"bstick" (broken stick).}

\item{check_arguments}{If TRUE, the function arguments are verified.
Should be set to FALSE to save time when the arguments have been checked elsewhere.}
}
\value{
When applied to a \code{species_distribution}, the function returns a tibble with
the fitted parameters in columns: \code{mu} and \code{sigma} (lognormal),
\code{alpha} (log-series), \code{prob} (geometric) or \code{max} (the number of individuals
of the most abundant species in a broken-stick distribution,
which has no parameter).

When applied to a numeric vector, the function returns a list of two tibbles.
The first one is \verb{$rac}:
\itemize{
\item \code{rank}: the rank of species, in decreasing order of abundance,
\item \code{abundance} : the theoretical number of individuals of the species.
}

The other one is \verb{$parameters} with the fitted parameters.
}
\description{
Fit a well-known distribution to a species distribution.
}
