% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacommunity.R
\name{metacommunity}
\alias{metacommunity}
\title{Aggregate communities into a metacommunity}
\usage{
metacommunity(
  abundances,
  name = "metacommunity",
  as_numeric = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{abundances}{An object of class \link{abundances} that contains several communities.}

\item{name}{The name of the metacommunity}

\item{as_numeric}{If \code{TRUE}, a number or a numeric vector is returned rather than a tibble.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\value{
An object of class \link{abundances} with a single row.
}
\description{
Aggregate communities into a metacommunity
}
\examples{
metacommunity(paracou_6_abd)
}
