% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{metacommunity}
\alias{metacommunity}
\title{Aggregate communities into a metacommunity}
\usage{
metacommunity(abundances, name = "metacommunity", check_arguments = TRUE)
}
\arguments{
\item{abundances}{An object of class "abundances" that contains several communities.}

\item{name}{The name of the metacommunity}
}
\value{
An object of class "abundances" with a single row.
}
\description{
Aggregate communities into a metacommunity
}
\examples{
metacommunity(paracou_6_abd)
}
