% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{mock_3sp}
\alias{mock_3sp}
\alias{mock_3sp_abd}
\alias{mock_3sp_dist}
\alias{mock_3sp_tree}
\title{Mock data}
\format{
\code{mock_3sp_abd} is a vector, \code{mock_3sp_dist} a matrix and
\code{mock_3sp_tree} an object of class \link[ape:phylo]{ape::phylo}.

An object of class \code{dist} of length 3.

An object of class \code{phylo} of length 4.
}
\usage{
mock_3sp_abd

mock_3sp_dist

mock_3sp_tree
}
\description{
A simple dataset to test diversity functions.
It contains 3 species with their abundances, their distance matrix and
their phylogenetic tree.
}
\examples{
mock_3sp_abd
mock_3sp_dist
plot(mock_3sp_tree)
axis(2)

}
\keyword{datasets}
