% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{mock_3sp}
\alias{mock_3sp}
\alias{mock_3sp_abd}
\alias{mock_3sp_dist}
\alias{mock_3sp_tree}
\title{Mock data}
\format{
An object of class \code{integer} of length 3.

An object of class \code{dist} of length 3.

An object of class \code{phylo} of length 4.
}
\usage{
mock_3sp_abd

mock_3sp_dist

mock_3sp_tree
}
\description{
A simple dataset to test diversity functions.
}
\details{
\code{mock_3sp_abd} is a vector of abundance, \code{mock_3sp_dist} their distance
matrix and \code{mock_3sp_tree} their phylogenetic tree.
}
\examples{
mock_3sp_abd
mock_3sp_dist
plot(mock_3sp_tree)
axis(2)

}
\keyword{datasets}
