% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo_divent.R
\name{plot.phylo_divent}
\alias{plot.phylo_divent}
\title{Plot phylo_divent Objects}
\usage{
\method{plot}{phylo_divent}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{Arguments passed to \link[stats:dendrogram]{stats::plot.dendrogram}.}
}
\description{
Plot objects of class "phylo_divent" produced by \link{as_phylo_divent}, that are
phylogenetic trees.
}
\examples{
# Paracou plot 6 species taxonomy
tree <- as_phylo_divent(paracou_6_taxo)
plot(tree, type="fan", show.tip.label=FALSE)

}
