% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
<<<<<<< HEAD
=======
  \item{base}{\code{\link[base]{plot}}}

>>>>>>> e57985df2e52c1426574e4899e7877ca73bab8ac
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

