% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R, R/species_distribution.R
\name{probabilities.numeric}
\alias{probabilities.numeric}
\alias{probabilities.abundances}
\alias{species_distribution}
\alias{as_species_distribution}
\alias{as_species_distribution.numeric}
\alias{as_species_distribution.matrix}
\alias{as_species_distribution.data.frame}
\alias{is_species_distribution}
\alias{plot.species_distribution}
\alias{as_probabilities}
\alias{as_probabilities.numeric}
\alias{as_probabilities.matrix}
\alias{as_probabilities.data.frame}
\alias{is_probabilities}
\alias{abundances}
\alias{as_abundances}
\alias{as_abundances.numeric}
\alias{as_abundances.matrix}
\alias{as_abundances.data.frame}
\alias{is_abundances}
\title{Species Distributions}
\usage{
\method{probabilities}{numeric}(
  x,
  estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  q = 0,
  ...,
  check_arguments = TRUE
)

\method{probabilities}{abundances}(
  x,
  estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  q = 0,
  ...,
  check_arguments = TRUE
)

species_distribution(x, names = NULL, weights = NULL, check_arguments = TRUE)

as_species_distribution(x, ...)

\method{as_species_distribution}{numeric}(x, ..., check_arguments = TRUE)

\method{as_species_distribution}{matrix}(
  x,
  names = NULL,
  weights = NULL,
  ...,
  check_arguments = TRUE
)

\method{as_species_distribution}{data.frame}(x, ..., check_arguments = TRUE)

is_species_distribution(x)

\method{plot}{species_distribution}(
  x,
  ...,
  distribution = c("lnorm", "lseries", "geom", "bstick"),
  log = "y",
  main = NULL,
  xlab = "Rank",
  ylab = NULL
)

as_probabilities(x, ...)

\method{as_probabilities}{numeric}(x, ..., check_arguments = TRUE)

\method{as_probabilities}{matrix}(x, names = NULL, weights = NULL, ..., check_arguments = TRUE)

\method{as_probabilities}{data.frame}(x, ..., check_arguments = TRUE)

is_probabilities(x)

abundances(
  x,
  round = TRUE,
  names = NULL,
  weights = NULL,
  check_arguments = TRUE
)

as_abundances(x, ...)

\method{as_abundances}{numeric}(x, round = TRUE, ..., check_arguments = TRUE)

\method{as_abundances}{matrix}(
  x,
  round = TRUE,
  names = NULL,
  weights = NULL,
  ...,
  check_arguments = TRUE
)

\method{as_abundances}{data.frame}(x, ..., check_arguments = TRUE)

is_abundances(x)
}
\arguments{
\item{x}{An object.}

\item{estimator}{One of the estimators of a probability distribution:
"naive" (the default value), or "Chao2013", "Chao2015", "ChaoShen" to estimate
the probabilities of the observed species in the asymptotic distribution.}

\item{unveiling}{One of the possible unveiling methods to estimate the probabilities
of the unobserved species: "none" (default, no species is added), "uniform"
(all unobserved species have the same probability) or "geometric" (the
unobserved species distribution is geometric).}

\item{richness_estimator}{An estimator of richness to evaluate the total number
of species. "jackknife" is the default value.
An alternative is "rarefy" to estimate the number of species such that the
entropy of the asymptotic distribution rarefied to the observed sample size equals
the actual entropy of the data.}

\item{jack_max}{The highest jackknife order allowed. Default is 10.
Allowed values are between 1 and 10.}

\item{coverage_estimator}{An estimator of sample coverage used by \link{coverage}:
"ZhangHuang" (the default value), \code{"Chao"}, \code{"Turing"} or \code{"Good"}.}

\item{q}{The order of diversity. Default is 0 for richness.
Used only to estimate asymptotic probability distributions when argument
\code{richness_estimator} is "rarefy". Then, the number of unobserved species is
fitted so that the entropy of order q of the asymptotic probability distribution
at the observed sample size equals the actual entropy of the data.}

\item{...}{Additional arguments to be passed to \link{plot}. Unused elsewhere.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{names}{The names of the species distributions.}

\item{weights}{The weights of the sites of the species distributions.}

\item{round}{If \code{TRUE}, the values of \code{x} are converted to integers.}
}
\description{
A Species Distribution is a \link[tibble:tibble]{tibble::tibble} containing species abundances or probabilities.
}
\details{
\code{species_distribution} objects include \code{abundances} and \code{probabilities} objects.

\code{as_species_distribution()}, \code{as_abundances()}  and \code{as_probabilities} format
the numeric, matrix or dataframe \code{x} so that appropriate
versions of community functions (generic methods such as \link{plot} or
\link{div_richness}) are applied.
Abundance values are rounded (by default) to the nearest integer.

\code{as_probabilities()} normalizes the vector \code{x} so that it sums to 1. It gives
the same output as \code{probabilities()} with \code{estimator = "naive"}.

TODO: These functions can be applied to data frames to calculate the joint diversity \insertCite{Gregorius2010}{divent}.

\code{species_distribution} objects objects can be plotted by \code{plot()} and \code{autoplot()}.
}
\examples{
# Paracou data
is_species_distribution(paracou_6_abd)
# Whittaker plot fitted by a log-normal distribution
autoplot(paracou_6_abd[1,], distribution = "lognormal")
}
\references{
\insertAllCited{}
}
