% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_distribution.R
\name{species_distribution}
\alias{species_distribution}
\alias{as_species_distribution}
\alias{as_species_distribution.numeric}
\alias{as_species_distribution.integer}
\alias{as_species_distribution.matrix}
\alias{as_species_distribution.data.frame}
\alias{is_species_distribution}
\alias{as_probabilities}
\alias{as_probabilities.numeric}
\alias{as_probabilities.integer}
\alias{is_probabilities}
\alias{as_abundances}
\alias{as_abundances.numeric}
\alias{as_abundances.integer}
\alias{as_abundances.matrix}
\alias{is_abundances}
\title{Species Distributions}
\usage{
as_species_distribution(x, ...)

\method{as_species_distribution}{numeric}(x, name = "", ..., check_arguments = TRUE)

\method{as_species_distribution}{integer}(x, name = "", ..., check_arguments = TRUE)

\method{as_species_distribution}{matrix}(x, names = NULL, ..., check_arguments = TRUE)

\method{as_species_distribution}{data.frame}(x, ..., check_arguments = TRUE)

is_species_distribution(x)

as_probabilities(x, ...)

\method{as_probabilities}{numeric}(
  x,
  estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "rarefy"),
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  q = 0,
  ...,
  check_arguments = TRUE
)

\method{as_probabilities}{integer}(
  x,
  estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "rarefy"),
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  q = 0,
  ...,
  check_arguments = TRUE
)

is_probabilities(x)

as_abundances(x, ...)

\method{as_abundances}{numeric}(x, name = "", round = TRUE, ..., check_arguments = TRUE)

\method{as_abundances}{integer}(x, name = "", ..., check_arguments = TRUE)

\method{as_abundances}{matrix}(x, names = "", round = TRUE, ..., check_arguments = TRUE)

is_abundances(x)
}
\arguments{
\item{x}{}

\item{...}{Additional arguments to be passed to \link{plot}. Unused elsewhere.}

\item{name}{The name of the species distribution.}

\item{check_arguments}{If \code{TRUE}, check the arguments of the function.}

\item{estimator}{One of the estimators of a probability distribution:
"naive" (the default value), or "Chao2013", "Chao2015", "ChaoShen" to estimate
the probabilities of the observed species in the asymptotic distribution.}

\item{unveiling}{One of the possible unveiling methods to estimate the probabilities
of the unobserved species: "none" (default, no species is added), "uniform"
(all unobserved species have the same probability) or "geometric" (the
unobserved species distribution is geometric).}

\item{richness_estimator}{An estimator of richness to evaluate the total number
of species. "jackknife" is the default value.
An alternative is "rarefy" to estimate the number of species such that the
entropy of the asymptotic distribution rarefied to the observed sample size equals
the actual entropy of the data.}

\item{jack_max}{The highest jackknife order allowed. Default is 10.
Allowed values are between 1 and 10.}

\item{coverage_estimator}{An estimator of sample coverage used by \link{coverage}:
"ZhangHuang" (the default value), \code{"Chao"}, \code{"Turing"} or \code{"Good"}.}

\item{q}{The order of diversity. Default is 0 for richness.
Used only to estimate asymptotic probability distributions when argument
\code{richness_estimator} is "rarefy". Then, the number of unobserved species is
fitted so that the entropy of order q of the asymptotic probability distribution
at the observed sample size equals the actual entropy of the data.}
}
\description{
A Species Distribution is a \link[tibble:tibble]{tibble::tibble} containing species abundances or probabilities.
}
\details{
\code{species_distribution} objects include \code{abundances} and \code{probabilities} objects.

\code{as_abundances()} formats the numeric or integer \code{x} so that appropriate versions of community functions (generic methods such as Diversity) are applied. Abundance values are rounded (by default) to the nearest integer.

\code{as_probabilities()} normalizes the vector \code{x} so that it sums to 1.
If the \code{estimator} is not "naive", the observed abundance distribution is used
to estimate the actual species distribution. The list of species will be changed:
zero-abundance species will be cleared, and some unobserved species will be added.
First, observed species probabilities are estimated folllowing
\insertCite{Chao2003;textual}{divent}, i.e. input probabilities are multiplied by
the sample coverage, or according to more sophisticated models:
\insertCite{Chao2013;textual}{divent}, a single-parameter model, or
\insertCite{Chao2015;textual}{divent}, a two-parameter model.
The total probability of observed species equals the sample coverage.
Then, the distribution of unobserved species can be unveiled: their number
is estimated according to the \code{richness_estimator}.
The coverage deficit (1 minus the sample coverage) is shared by the unobserved
species equally (\code{unveiling} = "uniform", \insertCite{Chao2013}{divent}) or
according to a geometric distribution (\code{unveiling} = "geometric", \insertCite{Chao2015}{divent}).

TODO: These functions can be applied to data frames to calculate the joint diversity \insertCite{Gregorius2010}{divent}.

\code{species_distribution} objects objects can be plotted by \code{plot} and \code{autoplot()}.
}
\examples{
# Paracou data TODO
is_species_distribution(asm_paracou_6)
# Whittaker plot fitted by a log-normal distribution
autoplot(asm_paracou_6, distribution = "lognormal")
}
\references{
\insertAllCited{}
}
