% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_distribution.R
\name{species_distribution}
\alias{species_distribution}
\alias{as_species_distribution}
\alias{as_species_distribution.numeric}
\alias{as_species_distribution.matrix}
\alias{as_species_distribution.data.frame}
\alias{is_species_distribution}
\alias{plot.species_distribution}
\alias{autoplot.species_distribution}
\alias{as_probabilities}
\alias{as_probabilities.numeric}
\alias{as_probabilities.matrix}
\alias{as_probabilities.data.frame}
\alias{is_probabilities}
\alias{abundances}
\alias{as_abundances}
\alias{as_abundances.numeric}
\alias{as_abundances.matrix}
\alias{as_abundances.data.frame}
\alias{is_abundances}
\title{Species Distributions}
\usage{
species_distribution(x, names = NULL, weights = NULL, check_arguments = TRUE)

as_species_distribution(x, ...)

\method{as_species_distribution}{numeric}(x, ..., check_arguments = TRUE)

\method{as_species_distribution}{matrix}(
  x,
  names = NULL,
  weights = NULL,
  ...,
  check_arguments = TRUE
)

\method{as_species_distribution}{data.frame}(x, ..., check_arguments = TRUE)

is_species_distribution(x)

\method{plot}{species_distribution}(
  x,
  ...,
  fit_rac = FALSE,
  distribution = c("lnorm", "lseries", "geom", "bstick"),
  ylog = "y",
  main = NULL,
  xlab = "Rank",
  ylab = NULL,
  palette = "Set1"
)

\method{autoplot}{species_distribution}(
  object,
  ...,
  fit_rac = FALSE,
  distribution = c("lnorm", "lseries", "geom", "bstick"),
  ylog = TRUE,
  main = NULL,
  xlab = "Rank",
  ylab = NULL,
  pch = ggplot2::GeomPoint$default_aes$shape,
  cex = ggplot2::GeomPoint$default_aes$size
)

as_probabilities(x, ...)

\method{as_probabilities}{numeric}(x, ..., check_arguments = TRUE)

\method{as_probabilities}{matrix}(x, names = NULL, weights = NULL, ..., check_arguments = TRUE)

\method{as_probabilities}{data.frame}(x, ..., check_arguments = TRUE)

is_probabilities(x)

abundances(
  x,
  round = TRUE,
  names = NULL,
  weights = NULL,
  check_arguments = TRUE
)

as_abundances(x, ...)

\method{as_abundances}{numeric}(x, round = TRUE, ..., check_arguments = TRUE)

\method{as_abundances}{matrix}(
  x,
  round = TRUE,
  names = NULL,
  weights = NULL,
  ...,
  check_arguments = TRUE
)

\method{as_abundances}{data.frame}(x, ..., check_arguments = TRUE)

is_abundances(x)
}
\arguments{
\item{x}{An object.}

\item{names}{The names of the species distributions.}

\item{weights}{The weights of the sites of the species distributions.}

\item{check_arguments}{If \code{TRUE}, the function arguments are verified.
Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{...}{Additional arguments to be passed to \link{plot}. Unused elsewhere.}

\item{fit_rac}{If \code{TRUE}, estimate a theoretical distribution and fit the data with it.}

\item{distribution}{The distribution of species abundances.
May be "lnorm" (log-normal), "lseries" (log-series), "geom" (geometric) or
"bstick" (broken stick).}

\item{ylog}{If \code{TRUE}, the Y-axis is in log-scale.}

\item{main}{The title of the plot.}

\item{xlab}{The label of the X-axis.}

\item{ylab}{The  label of the Y-axis.}

\item{palette}{The name of a color palette, recognized by \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal}.}

\item{object}{An object of class \link{species_distribution}.}

\item{pch}{The plotting characters. See \link[graphics:points]{graphics::points}.}

\item{cex}{The character expansion (size) of the points. See \link[graphics:points]{graphics::points}.}

\item{round}{If \code{TRUE}, the values of \code{x} are converted to integers.}
}
\description{
A Species Distribution is a \link[tibble:tibble]{tibble::tibble} containing species abundances or probabilities.
}
\details{
\code{species_distribution} objects include \code{abundances} and \code{probabilities} objects.

\code{as_species_distribution()}, \code{as_abundances()}  and \code{as_probabilities} format
the numeric, matrix or dataframe \code{x} so that appropriate
versions of community functions (generic methods such as \link{plot} or
\link{div_richness}) are applied.
Abundance values are rounded (by default) to the nearest integer.

\code{as_probabilities()} normalizes the vector \code{x} so that it sums to 1. It gives
the same output as \code{probabilities()} with \code{estimator = "naive"}.

TODO: These functions can be applied to data frames to calculate the joint diversity \insertCite{Gregorius2010}{divent}.

\code{species_distribution} objects objects can be plotted by \code{plot()} and \code{autoplot()}.
}
\examples{
# Paracou data
is_species_distribution(paracou_6_abd)
# Whittaker plot fitted by a log-normal distribution
autoplot(paracou_6_abd[1,], distribution = "lnorm")
}
\references{
\insertAllCited{}
}
